backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.4
oauthlib==3.1.0
Pillow==6.2.0
pkg-resources==0.0.0
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
